coord,sql: support CREATE TEMPORARY MATERIALIZED VIEW (#5542)
This change allows users to create temporary materialized views in their
Materialize instance--thereby creating temporary indexes, as well. Temporary
materialized views (and indexes) behave identically to other temporary
objects: they are only visible to the user that creates them, and they are
automatically dropped when the user's connection is dropped or the
Materialize instance is stopped.