sql: move all WITH option normalization into normalize mod
Both statement purification and planning need to normalize WITH options,
but some of the WITH option normalization code lived in the plan module.
Rather than reaching into the plan module from the normalize module to
pull it out, move that WITH option code into the normalize module, which
is naturally depended upon by both the plan and pure modules.

This commit is pure code movement.