fixup! kgen: avoid allocation when generating key data (#5572)
The key variable was no longer being assigned by the if / else statements but
it was being used. This resulted in us writing a null key to Kafka. Remove this
variable, as it's no longer necessary. This then makes it obvious that key_buf
is the thing we should be using for the key.