coord: reimplement cancel
Re-implement how cancellation works by sharing a tokio::watch channel with
a connection and the coordinator. The coordinator can send a cancelled
message on it that the connection can listen for when it wants to. This
allows us to add cancellation points anywhere in the connection handler
instead of depending on shutting down dataflows to do that. Currently
this can be used to shut down a running TAIL or FETCH. (SELECT remains
unchanged: it can be cancelled as long as the peek is pending.)

Npgsql v4 test removed because the exception type changed for cancellation
and it wasn't worth figuring out how to have a single file support
both versions.

Fixes #5242