Fixup username in materialize connection strings (#5586)
Continued from #5555

Update chbenchmark build image to an LTS Ubuntu image

Eoan is no longer supported, migrate the image to an LTS supported
version. I originally did this as a multistage build but it was overly
simplistic and splitting the image doesn't buy us much. Many of the ODBC
dependencies need to be installed in the second image.