materialized: require explicit client construction in tests
Change util::start_server so that it does not return a client that is
connected to the server. Tests must now explicitly call `server.connect`
to get a client with an acceptable TLS configuration.

This change is required for forthcoming changes to allow servers that
*require* TLS, in which case trying to automatically return a connected
client from util::start_server is problematic, as start_server does not
know what TLS configuration to use for the client.