Add a medium sized benchmark for running on laptops (#5628)
This benchmark runs a much smaller dataset that results in peak memory
utilization of about 10GB and should be a good benchmark to run on a
Macbook Pro.

All of the schemas are the same, use the same file regardless of
distribution.

Configure 8 workers and 8 generator threads in benchmark medium.

Complain if the number of threads does not divide evenly into the number
of records to write.