doc/user: rework table overview and restrictions
Rework the CREATE TABLE doc to change the messaging of tables around. In
particular, this commit removes any references to tables as
"non-streaming data sources", as in Materialize a core tenent is that a
table is just a realization of a stream at a particular moment in time.

Also stress that most users should use sources in almost all cases, and
explicitly explain the pitfalls of tables, particularly around memory
usage.

Related to #5611 and #5627.