Add mzbench script for running benchmarks multiple times (#5686)
The mzbench script is documented under doc/developer/mzbench.md. It's
purpose is to run a benchmark many times, varying parameters such as git
checkout and number of materialized worker threads. mzbench is aware of
how many CPUs are on the host system and tries to pick interesting worker
counts accordingly.

This requires rewriting a bit of the workflows, as they must now be broken
into two parts: setup and run.