materialized: tweak panic message some more
Expand the panic message out to multiple lines, so the various fields
are more clear. The effect is that a panic that previously rendered like
so

    materialized v0.7.1-dev (4ffb6910a) listening on 0.0.0.0:6875...
    2021-02-11T02:59:38.549439Z ERROR panic: <unnamed>: something is wrong at '/home/benesch/Sites/materialize/materialize/src/coord/src/coord.rs:373:9'
       0: materialized::handle_panic
                 at src/materialized/src/bin/materialized/main.rs:575:78
       1: core::ops::function::Fn::call
    ...

now renders as:

    2021-02-11T03:14:16.915980Z ERROR panic: something is wrong
    thread: <unnamed>
    location: /home/benesch/Sites/materialize/materialize/src/coord/src/coord.rs:373:9
    version: 0.7.1-dev (4ffb6910a5deccede61abd9cb67c8a29fe38145c)
    backtrace:
       0: materialized::handle_panic
                 at src/materialized/src/bin/materialized/main.rs:588:21
       1: core::ops::function::Fn::call
    ...

I think this makes it much more clear that "<unnamed>" refers to a
thread.

The other big win of the new format is the inclusion of the version
information. Empirically we've found users are good at getting us the
backtrace, but not always the whole log file with the version
information. Reproducing the version info in the panic handler should
help with that.