coord: clean up the public interface
Add some new docstrings and copyedit some existing ones. Remove some
functions and types that didn't need to be public. Tweak the output of
`serve` so that it returns an opaque handle object and a client; this
makes booting a coordinator less grotty. Previously you had to thread
some MPSC transmitters and receivers in just the right ways.