[Orthogonality] Split sources into "bare" and "transformed" versions (#5570)
We now split the concept of source into "bare source" and "source" (or: "transformed" source, "derived" source, etc.). The intent is that the bare source should represent the raw data from the connector, and the transformed version represents the full richness of what we have been calling a "source" until now, including decoding, Debezium processing, upsert envelope application, and so on.

In this PR, the only thing we move out of the bare source is the "explode" call on the diff column, so the bare source still involves relatively a lot of complex logic. But it's a start!

Next work:

    Move more stuff out of the bare source and into the derived source; in particular, upsert envelope and avro/csv/etc. decoding.
    Expose the bare source and transformations on it to the user, in SQL (allowing custom transformations)
    Move timestamping to happen at the end of the derived source, rather than being a property of the bare source. This will make the end of source processing a clean boundary between the ingest pipeline and the rest of the system.