s3 sqs: More aggresively release SQS messages when clients shut down
Every time an SQS client reads a message from SQS that messages is marked
invisible, and other clients that read from that queue will not see the
messages on poll until the visibility timeout[1]. The default timeout is 5
minutes, but it can be extended up to 12 hours.

The primary reason to do this is to improve the existing test reliability, but
it will also make interactions with sqs queues a little more intuitive in edge
cases that involve dropping sources during exploratory SQL programming.

[1]: https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-visibility-timeout.html