include broker configuration when we validate the kafka config
During purification we do a shallow validation of the kafka config to
suss out any major issues with authentication early on.
Previously we didn't include the bootstrap.servers parameter when we did
this validation, resulting in a logged warning like:

```
INFO librdkafka: CONFWARN [thrd:app]: No bootstrap.servers configured: client will not be able to connect to Kafka cluster
```

This warning was benign since the config is never used to build a kafka
client, but it is extremely misleading when trying to debug unrelated
kafka issues.

This change includes the bootstrap.servers config parameter to remove
the misleading warning.