Record benchmark metrics to Kafka (#5792)
This PR updates the wait_metric script to write the results of the
benchmarks directly to a Kafka topic, in addition to printing the
results to stdout. When the benchmarks are run in our AWS
account, they will be configured to write to a central Kafka instance.

When run locally, they are configured to write metrics back to the
Kafka instance used for the tests themselves. The goal is to have
the benchmark results stored in a central, persistent location
when they are run in the cloud.