Simplify avro-upsert mzcompose by using env substitution (#6014)
Now that we can substitute environment variables into the definition of
sub-workflows, we can eliminate much of the duplication the avro-upsert
benchmark file. This change allows for benchmark-large,
benchmark-medium and benchmark-ci to share the same workflow
definition, but with different parameters.

Also, rename `benchmark` to `benchmark-large` so that all 3 benchmark
names are consistent.