Increase inline storage of Row without changing inline size (#6026)
SmallVec allows you to specify how much inline storage it should use before heap allocating. The
`16` bytes value there seemed suspicious to me because it doesn't seem like the right number of free
bytes necessary for any particular shape of "is overflowed" flag. Indeed, bumping it up to 23
doesn't change the size from 32 bytes, the test passes with and without this change, but fails if I
bump it up to 24.