dataflow: remove the buffer_message trait from SourceInfo
Note that previously, Kafka sources were somewhat "intelligent" in that they could
detect when the read cursor for a partition was past matching a potential window
of [lower, upper) offsets for a timestamp t and thus would automatically close t
for that partition. Kafka sources no longer do that, and have to observe (aka, be
able to timestamp a message at an offset >= upper) before the timestamp gets closed.

This should have no effect on the current implementation of RT sources, and it affects
BYO sources in that the consistency topic has to provide some timestamp information
at an offset x before timestamps assigned to offsets < x can be closed. This change
required also changing a testdrive test to provide some additional timestamping information
before selecting from a materialized source.