*: make Kinesis sources present as single-partition
Kinesis does parallelism with shards which are similar to Kafka partitions but
with fewer guarantees. For now, all Kinesis shards for a source instance are read
by the same worker. However, we were still attempting to register the set of shards
and the maximum "offset" read up to in each shard even though that information
wasn't being used for anything and more importantly, it required that we potentially
write to `ConsistencyInfo` when we read from a Kinesis source.

This commit changes Kinesis partition ID's to be singletons like File or S3
partition ids to simplify the behavior. The underlying source reader still
occaisonally checks for new shards, but this information is private and does
not need to be shared with the rest of the system.

Note that even without this change the offsets we get for various Kinesis messages
are entirely nondeterministic across restarts.