repr,expr,sql: permit jsonb numbers to be i64 or f64
Previously we rejected JSON numbers that were integers greater than
2^53-1, as we stored JSON numbers as double-precision floats. (Integers
greater than 2^53-1 are not guaranteed to be exactly representable by a
double-precision float, and we didn't want to silently change data.)

Our users, however, expect integers up to at least 2^64-1 to be
supported. This commit facilitates that by allowing numbers in the jsonb
type to be *either* an i64 or an f64.

We're still not quite PostgreSQL compatible, since PostgreSQL allows any
number that will fit in its numeric type, which supports up to 1000
digits of precision. But this commit brings us a good deal closer to
PostgreSQL compatibility, and should be sufficient for most real-world
uses of Materialize.

This commit also cleans up casting *from* jsonb to other types.
Previously if those casts failed they'd return NULL; now they return an
error to match PostgreSQL.

Fix #5919.