dataflow: reorganize how we use timestamping information
This commit teaches BYO sources to keep a copy of the timestamp binding that
satisfies the most recent offset's timestamping needs and try to reuse it for
as long as possible. This simplifies the BYO timestamp assignment logic, but
adds an assumption that timestamp bindings will correspond to contiguous offset
intervals with no gaps or out of order-ness.