coord,sql: Use the --timestamp-frequency option to control source timestamp frequency
By default all CREATE SOURCES are created with a timestamp frequency
of 1s. In order to speed up tests, use the existing --timestamp-frequency
hidden option to specify a custom default timestamp frequency.

This option was previously used to control the frequency of the timestamper thread.
In order to switch it to its new use, a new default of 1s is now provided.
The previous default of 10ms that previously applied to the timestamper
thread is now hard-coded when creating the Timestamper object in src/coord/src/coord.rs