Scrape prometheus registry & feed into built-in tables
This adds a thread to the coordinator that bumps the metrics over to
a the logging worker via a broadcast message, then adds them to a set of
builtin tables (with appropriate expiry). This should allow our users
to see how the system is behaving from the psql commandline (as well
as allow our benchmark tasks to make decisions without spinning a bunch
of containers).