doc/design: add a volatile source proposal
Our users want to connect Materialize to upstream sources that do not provide
guarantees about what data they will present. These sources are in tension with
the goal of providing a deterministic, replayable system.

This design doc proposes to:

  * Introduce "volatile source" as a term of art that describes a source that
    does not have upstream durability.

  * Use the terminology consistently in the docs site and error messages to
    educate our users about volatile sources.

  * Restrict the use of volatile sources in contexts where they could impede
    future improvements to the system's determinism and correctness.

The top-line goal is to unblock PubNub and SSE sources for use with the
cloud product.