Propose clear definitions of our log levels
This mostly corresponds to the way that we use log statements currently, but it is also a proposal
for making this understanding normative.

Log levels are always going to be a little bit vague, with unfortunately-broad lines between
different levels. However, logs are a useful line of defense in monitoring, and having clear
standards for how we argue whether something should be logged at a specific level can help ensure
that our on-disk logs are useful, instead of just being noise.

If we agree that these are good levels, I'll do another audit of our log statements to help ensure
that we do match these levels.