materialized: downgrade unknown protocol log message to debug
Empirically lots of things will try to connect to a running materialized
server that do not speak HTTP or pgwire. These connections do not rise
to the level of error, warn, or even info; hide them away in debug logs.
This should stop the steady stream of user concern we get about this log
message.

Motivated by #6311.