coord: make catalog responsible for generating builtin table updates
Push the responsibility for generating the updates to builtin tables
into the catalog. Previously, a catalog transaction would generate a
list of events ("created item", "dropped item", etc.) that the
coordinator would translate into builtin table updates.

The big downside with the old approach was that after a drop operation,
the catalog reflects the *new* state of the system while the coordinator
needs to emit a record reflecting the *old* state of the system. This is
ok when an event is self contained ("drop database"), but this is not
the case for index updates. Index updates include information not just
about the index but about the object upon which the index is built. When
generating the builtin table update for a "drop index" operation, the
coordinator really wants to be able to look up the metadata for the
object upon which the index is built in the catalog; but if the index
was dropped along with the view upon which it was built in the same
transaction, then that view won't be available in the catalog. We'd
hacked our way around this problem for indexes, but the approach doesn't
generalize to the forthcoming volatility bit.

In the new approach, the catalog generates the updates for a drop
operation *before* applying the operation, while it generates the
updates for a create operation *after* applying the operation. That
neatly sidesteps the problem described above. When generating the update
for a "drop index" operation, the index has not yet been dropped, and so
it is safe to look up whatever metadata is necessary in the catalog.