doc/design: document postgres sources design
Many users are keen to use materialize with a dataset existing in a
postgres upstream database. The current best-practice to ingest the data
in realtime is to setup a Kafka cluster and use Debezium to convert the
postgres CDC format into Kafka topics that can then be consumed by
materialize.

This design doc proposes to add a new type of source called `POSTGRES`
where materialized directly connects to an upstream database and
interprets the replication protocol, ingesting row updates in real time.

Users should be able to try materialize out simply by running the
materialize binary and pointing it to their postgres database, no extra
infrastructure needed.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>